% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess.rank}
\alias{assess.rank}
\title{Ranking assess funds using returns xts dataset}
\usage{
\method{assess}{rank}(returns.xts, choice.stock, step = F, ...)
}
\arguments{
\item{returns.xts}{an \code{xts} object about returns dataset}

\item{method}{returns mean or sd or Sharpe ratio}
}
\value{
data.frame type about assessment result
}
\description{
Ranking assess funds using returns xts dataset & rank plotting
}
\examples{
index <- matrix(rnorm(90,1000, 8^2), 30, 3)
data <- xts(round(index, 1), as.Date(16001:16030))
data[9:13, 3] <- NA
names(data) <- paste("Fund", 1:3, sep = "_")
returns <- Xday_returns(data, 1)
}

