% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_returns.R
\name{assess_returns}
\alias{assess_returns}
\title{Assess funds using returns xts dataset}
\usage{
assess_returns(returns.xts, method = c("mean", "sd", "S.R"), ...)
}
\description{
Assess funds using returns xts dataset
}
\examples{
index <- matrix(rnorm(60,1000, 8^2), 20, 3)
data <- xts(round(index, 1), as.Date(16001:16020)) # xts package for time series data
names(data) <- paste("Fund", 1:3, sep="_")
returns <- Xday_returns(data, 1)

assess_returns(returns, "mean")
assess_returns(returns, "sd")
assess_returns(returns, "S.R")
}

