% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{descr}
\alias{descr}
\title{Describe dataset using psych::describe() function}
\usage{
descr(x, choice = c("vars", "n", "mean", "sd", "median", "min", "max",
  "range"), ...)
}
\arguments{
\item{x}{A data frame or matrix}

\item{...}{Arguments to be passed to \code{psych::describe()}}
}
\description{
Alias function of \code{psych::describe()}
}
\examples{
desc(mtcars)
desc(mtcars, boxplot = T)
bound_trim(c(1, 5, 9, 10), 0, 9)
bound_trim(rnorm(100), -1.96, 1.96)
}

