% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ky_colors.R
\name{ky_colors}
\alias{ky_colors}
\title{Color Palettes}
\usage{
ky_colors(n, alpha = 1)
}
\arguments{
\item{n}{the number of colors (<U+FFFD><U+FFFD> 1) to be in the palette.}

\item{alpha}{the alpha transparency, a number in [0,1].}
}
\description{
Color Palettes for kyobo
}
\examples{
(colors1 <- ky_colors(9))
pie(rep(1, 9), col = colors1, clockwise = T)

(colors2 <- ky_colors(15, alpha = .5))
pie(rep(1, 15), col = colors2, clockwise = T)
}

