% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmadjust.R
\name{mmadjust}
\alias{mmadjust}
\title{Adjust min & max}
\usage{
mmadjust(x, min, max)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Adjust min & max about numeric vector
}
\examples{
x <- c(NA, 1:100*.01, NA)
mmadjust(x, .2, .8)
}

