% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plot.R
\name{tm1plot}
\alias{tm1plot}
\title{ts and moving average plot for single fund item}
\usage{
tm1plot(xts, choice.stock, mv = c(20, 60, 120), plotly = F)
}
\description{
ts and moving average plot for single fund item
}
\examples{
set.seed(1004)
index <- matrix(rnorm(3000, 1000, 5^2), 1000, 3)
data <- xts(round(index, 1), as.Date(16001:17000))
names(data) <- paste("Fund", 1:3, sep="_")
data

tm1plot(data, "Fund1")
tm1plot(data, "Fund1", plotly=T)
}

