% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plot.R
\name{tmplot}
\alias{tmplot}
\title{ts and moving average plot for all fund items}
\usage{
tmplot(xts, mv = c(20, 60, 120), plotly = F, ...)
}
\description{
ts and moving average plot for all fund items
}
\examples{
library("quantmod")

data <- getSymbols("^ks11", from=Sys.Date()-1000, to=Sys.Date(), auto.assign=F)

tmplot(data)
tmplot(data, plotly=T)
}

