% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trimming of whitespace in character}
\usage{
trim(data, method = "trailing")
}
\description{
Trimming of whitespace in character
}
\examples{
trim("  Wow!  ", method = "leading")
trim("  Wow!  ", method = "trailing")
}

