% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{xdiff_returns}
\alias{xdiff_returns}
\title{Accounting x diff returns}
\usage{
xdiff_returns(xts, x = 1, log.returns = F, start.omit = T)
}
\arguments{
\item{xts}{an \code{xts} object}

\item{x}{an integer value}

\item{log.returns}{logical asking whether to calculate log returns}
}
\value{
xts type about returns dataset
}
\description{
일괄적으로 모든 x 시차 값을 기준으로 수익률을 계산합니다.
}
\examples{
xdiff_returns(sample_index, 1)
xdiff_returns(sample_index, 1, log.returns = T)
}

