% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{xdiffs_returns}
\alias{xdiffs_returns}
\title{Accounting x diff returns}
\usage{
xdiffs_returns(xts, x, log.returns = F, start.omit = T)
}
\arguments{
\item{xts}{an \code{xts} object}

\item{x}{an integer vector}

\item{log.returns}{logical asking whether to calculate log returns}
}
\value{
xts type about returns dataset
}
\description{
각기 다른 x 시차 값을 기준으로 수익률을 계산합니다. 따라서 \code{x}는 \code{n = nrow(xts)} 만큼의 벡터타입값을 받아야 합니다.
}
\examples{
xdiffs_returns(sample_index, x = 0:19)
xdiffs_returns(sample_index, x = 0:19, log.returns = T)
}

