% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efff.R
\name{efff}
\alias{efff}
\title{Efficient Frontier}
\usage{
efff(returns, short = "no", max.allocation = NULL, risk.premium.up = 0.9,
  risk.increment = 1e-04, rg = NA, plot = T, rfr = 0)
}
\arguments{
\item{rg}{numeric value, returns of goal}

\item{plot}{logical value, return the plot?}
}
\description{
Efficient Frontier
}
\examples{
xdiff_returns(sample_index, x = 1) \%>\% efff(rg = .01, rfr = .001)
}

