% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efff.R
\name{efff}
\alias{efff}
\title{Efficient Frontier}
\usage{
efff(returns, rg = NA, rfr = 0, short = "no", max.allocation = NULL,
  risk.premium.up = 0.9, risk.increment = 1e-04, plot.only.efff = T,
  plotly = F)
}
\arguments{
\item{rg}{numeric value, returns of goal}
}
\description{
Efficient Frontier using excess returns
}
\examples{
returns <- xdiff_returns(sample_index, x = 1)
efff(returns, rg = .01, rfr = .001)
}

