% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvp.R
\name{mvp}
\alias{mvp}
\title{Minimum variance portfolio using returns xts dataset}
\usage{
mvp(returns.xts, precision = 0.01, rfr = 0, rg = NA, plot = T, ...)
}
\arguments{
\item{returns.xts}{an \code{xts} object about returns dataset}

\item{precision}{grid precision about investment weight}

\item{rfr}{risk free rate}

\item{rg}{goal of returns}
}
\value{
data.frame type about assessment result
}
\description{
Minimum variance portfolio & Market portfolio using returns xts dataset
}
\examples{
returns <- xdiff_returns(sample_index, 1)
mvp(returns, rfr = 0.001)
}

