% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plot.R
\name{tm1plot}
\alias{tm1plot}
\title{ts and moving average plot for single fund item}
\usage{
tm1plot(xts, choice.stock, mv = c(20, 60, 120), plotly = F, ...)
}
\description{
ts and moving average plot for single fund item
}
\examples{
library("quantmod")

data <- getSymbols("^ks11", from = Sys.Date() - 1000, to = Sys.Date(), auto.assign = F)

tm1plot(data, "KS11.Close", plotly = T)
}

