% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{Xday_returns}
\alias{Xday_returns}
\title{Account X day returns}
\usage{
Xday_returns(xts, Xday = 1L, log.returns = F)
}
\description{
Account X day returns
}
\examples{
Xday_returns(xts, 365, log.returns=T)
}

