% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{Xday_returns}
\alias{Xday_returns}
\title{Account X day returns}
\usage{
Xday_returns(data, X, start.column = 7)
}
\description{
Account X day returns
}
\examples{
Xday_returns(data, 365, start.column=7)
}

