% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{Xday_returns}
\alias{Xday_returns}
\title{Account X day returns}
\usage{
Xday_returns(xts, Xday = 1L, log.returns = F, start.omit = T)
}
\arguments{
\item{xts}{an \code{xts} object}

\item{Xday}{an integer value}

\item{log.returns}{logical asking whether to calculate log returns}
}
\value{
xts type about returns dataset
}
\description{
Account X day returns
}
\examples{
index <- matrix(rnorm(60,1000, 8^2), 20, 3)
data <- xts(round(index, 1), as.Date(16001:16020)) # xts package for time series data
names(data) <- paste("Fund", 1:3, sep = "_")
Xday_returns(data, 1)
Xday_returns(data, 1, log.returns = T)
}

